Resolves all known bugs. UI loads and renders correctly (although could still use some tweaking to look better). Handles cases where inventory or equipment are null/empty, including on login.
