+++ src/main/java/com/loadoutsaver/DataIO.java
@@ -23,17 +23,6 @@ public class DataIO {
 
     private DataIO() {}
 
-    /**
-     * Default loader that reads one loadout per line.
-     * @param inputs The stream containing loadouts, one per line.
-     * @throws IOException If reading from the stream fails.
-     */
-    public static List<ILoadout> Load(InputStream inputs) throws IOException {
-        byte[] loaded = inputs.readAllBytes();
-        String decoded = new String(loaded, DataIO.ENCODING);
-        return Parse(decoded);
-    }
-
     public static List<ILoadout> Parse(String decoded) {
         String[] lines = decoded.split("\n");
         List<ILoadout> result = Arrays.stream(lines).filter(
@@ -43,7 +32,7 @@ public class DataIO {
                     try {
                         return LoadoutImpl.Deserializer.DeserializeString(raw);
                     } catch (IllegalArgumentException iae) {
-                        System.err.println("Could not parse loadout: " + decoded);
+                        System.err.println("Could not parse loadout: " + raw);
                         return null;
                     }
                 }
@@ -60,16 +49,5 @@ public class DataIO {
     public static String FullSerialize(Collection<ILoadout> loadouts) {
         List<String> encoded = loadouts.stream().map(ISerializable::SerializeString).collect(Collectors.toList());
         return String.join("\n", encoded);
-    }
-
-    /**
-     * Default saver that saves loadouts on separate lines (when not encoded).
-     * @param outputs The stream to which outputs should be written.
-     * @param loadouts The loadouts to be saved.
-     * @throws IOException If writing to the stream fails.
-     */
-    public static void Save(OutputStream outputs, Collection<ILoadout> loadouts) throws IOException {
-        byte[] fullOutputs = FullSerialize(loadouts).getBytes(DataIO.ENCODING);
-        outputs.write(fullOutputs);
     }
 }
