+++ src/main/resources/com/loadoutsaver/ui/ammodefault.png
