+++ src/main/resources/com/loadoutsaver/ui/amuletdefault.png
