+++ src/main/resources/com/loadoutsaver/ui/bodydefault.png
