+++ src/main/resources/com/loadoutsaver/ui/bootsdefault.png
