+++ src/main/resources/com/loadoutsaver/ui/capedefault.png
