+++ src/main/resources/com/loadoutsaver/ui/emptyequipment.png
