+++ src/main/resources/com/loadoutsaver/ui/equipmentbackground.png
