+++ src/main/resources/com/loadoutsaver/ui/equipmentbackgroundgrid.png
