+++ src/main/resources/com/loadoutsaver/ui/glovesdefault.png
