+++ src/main/resources/com/loadoutsaver/ui/helmetdefault.png
