+++ src/main/resources/com/loadoutsaver/ui/inventorybackground.png
