+++ src/main/resources/com/loadoutsaver/ui/legsdefault.png
