+++ src/main/resources/com/loadoutsaver/ui/ringdefault.png
