+++ src/main/resources/com/loadoutsaver/ui/shielddefault.png
