+++ src/main/resources/com/loadoutsaver/ui/weapondefault.png
